+++ pkg/domainservice/validateDependenciesDomainUseCase_test.go
@@ -12,9 +12,9 @@ import (
 )
 
 var (
-	version1_0_0_1, _ = core.ParseVersion("1.0.0-1")
-	version2_0_0_1, _ = core.ParseVersion("2.0.0-1")
-	version2_0_0_3, _ = core.ParseVersion("2.0.0-3")
+	version1001, _ = core.ParseVersion("1.0.0-1")
+	version2001, _ = core.ParseVersion("2.0.0-1")
+	version2003, _ = core.ParseVersion("2.0.0-3")
 
 	officialNamespace   = common.DoguNamespace("official")
 	premiumNamespace    = common.DoguNamespace("premium")
@@ -33,6 +33,7 @@ var (
 	officialPlantuml      = common.QualifiedDoguName{Namespace: officialNamespace, SimpleName: common.SimpleDoguName("plantuml")}
 	officialUnknownDogu   = common.QualifiedDoguName{Namespace: officialNamespace, SimpleName: common.SimpleDoguName("unknownDogu")}
 	helloworldBluespice   = common.QualifiedDoguName{Namespace: helloworldNamespace, SimpleName: common.SimpleDoguName("bluespice")}
+	ldapMapper            = common.QualifiedDoguName{Namespace: officialNamespace, SimpleName: common.SimpleDoguName("ldap-mapper")}
 
 	ctx = context.Background()
 )
@@ -47,13 +48,13 @@ func Test_checkDependencyVersion(t *testing.T) {
 		args    args
 		wantErr bool
 	}{
-		{name: "exact version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_1}, expectedVersion: "2.0.0-1"}, wantErr: false},
-		{name: "has lower version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_1}, expectedVersion: ">=2.0.0-2"}, wantErr: true},
-		{name: "has higher version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_3}, expectedVersion: ">=2.0.0-2"}, wantErr: false},
-		{name: "needs lower version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_3}, expectedVersion: "<=2.0.0-2"}, wantErr: true},
-		{name: "needs higher version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_1}, expectedVersion: ">2.0.0-1"}, wantErr: true},
-		{name: "no constraint", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_1}, expectedVersion: ""}, wantErr: false},
-		{name: "not parsable expected version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2_0_0_1}, expectedVersion: "abc"}, wantErr: true},
+		{name: "exact version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2001}, expectedVersion: "2.0.0-1"}, wantErr: false},
+		{name: "has lower version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2001}, expectedVersion: ">=2.0.0-2"}, wantErr: true},
+		{name: "has higher version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2003}, expectedVersion: ">=2.0.0-2"}, wantErr: false},
+		{name: "needs lower version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2003}, expectedVersion: "<=2.0.0-2"}, wantErr: true},
+		{name: "needs higher version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2001}, expectedVersion: ">2.0.0-1"}, wantErr: true},
+		{name: "no constraint", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2001}, expectedVersion: ""}, wantErr: false},
+		{name: "not parsable expected version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: version2001}, expectedVersion: "abc"}, wantErr: true},
 		{name: "not parsable actual version", args: args{doguInBlueprint: domain.Dogu{Name: officialNginx, Version: core.Version{Raw: "abc"}}, expectedVersion: "2.0.0-1"}, wantErr: true},
 	}
 	for _, tt := range tests {
@@ -80,8 +81,8 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			args: args{
 				effectiveBlueprint: domain.EffectiveBlueprint{
 					Dogus: []domain.Dogu{
-						{Name: officialRedmine, Version: version1_0_0_1},
-						{Name: officialPostgres, Version: version1_0_0_1},
+						{Name: officialRedmine, Version: version1001},
+						{Name: officialPostgres, Version: version1001},
 					},
 				},
 			},
@@ -92,8 +93,8 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			args: args{
 				effectiveBlueprint: domain.EffectiveBlueprint{
 					Dogus: []domain.Dogu{
-						{Name: officialRedmine, Version: version1_0_0_1},
-						{Name: officialPostgres, Version: version1_0_0_1},
+						{Name: officialRedmine, Version: version1001},
+						{Name: officialPostgres, Version: version1001},
 					},
 				},
 			},
@@ -104,7 +105,7 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			args: args{
 				effectiveBlueprint: domain.EffectiveBlueprint{
 					Dogus: []domain.Dogu{
-						{Name: officialRedmine, Version: version1_0_0_1},
+						{Name: officialRedmine, Version: version1001},
 					},
 				},
 			},
@@ -116,7 +117,7 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			args: args{
 				effectiveBlueprint: domain.EffectiveBlueprint{
 					Dogus: []domain.Dogu{
-						{Name: officialRedmine2, Version: version1_0_0_1},
+						{Name: officialRedmine2, Version: version1001},
 					},
 				},
 			},
@@ -128,7 +129,7 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			args: args{
 				effectiveBlueprint: domain.EffectiveBlueprint{
 					Dogus: []domain.Dogu{
-						{Name: officialK8sCesControl, Version: version1_0_0_1},
+						{Name: officialK8sCesControl, Version: version1001},
 					},
 				},
 			},
@@ -138,7 +139,7 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			name: "missing nginx-static and nginx ingress on nginx dependency",
 			args: args{effectiveBlueprint: domain.EffectiveBlueprint{
 				Dogus: []domain.Dogu{
-					{Name: officialScm, Version: version1_0_0_1},
+					{Name: officialScm, Version: version1001},
 				},
 			}},
 			wantErr: true,
@@ -147,9 +148,18 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus(t *te
 			name: "ok with nginx dependency",
 			args: args{effectiveBlueprint: domain.EffectiveBlueprint{
 				Dogus: []domain.Dogu{
-					{Name: officialPlantuml, Version: version1_0_0_1},
-					{Name: k8sNginxStatic, Version: version1_0_0_1},
-					{Name: k8sNginxIngress, Version: version1_0_0_1},
+					{Name: officialPlantuml, Version: version1001},
+					{Name: k8sNginxStatic, Version: version1001},
+					{Name: k8sNginxIngress, Version: version1001},
+				},
+			}},
+			wantErr: false,
+		},
+		{
+			name: "registrator should be ignored",
+			args: args{effectiveBlueprint: domain.EffectiveBlueprint{
+				Dogus: []domain.Dogu{
+					{Name: ldapMapper, Version: version1001},
 				},
 			}},
 			wantErr: false,
@@ -178,7 +188,7 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus_NotFo
 	// when
 	err := useCase.ValidateDependenciesForAllDogus(ctx, domain.EffectiveBlueprint{
 		Dogus: []domain.Dogu{
-			{Name: officialUnknownDogu, Version: version1_0_0_1},
+			{Name: officialUnknownDogu, Version: version1001},
 		},
 	})
 	// then
@@ -207,8 +217,8 @@ func TestValidateDependenciesDomainUseCase_ValidateDependenciesForAllDogus_colle
 	// when
 	err := useCase.ValidateDependenciesForAllDogus(ctx, domain.EffectiveBlueprint{
 		Dogus: []domain.Dogu{
-			{Name: officialRedmine, Version: version1_0_0_1},
-			{Name: helloworldBluespice, Version: version1_0_0_1},
+			{Name: officialRedmine, Version: version1001},
+			{Name: helloworldBluespice, Version: version1001},
 		},
 	})
 	// then