+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.1] - 2024-03-20
+### Fixed
+- [#74] Fix ldap-mapper dependency check by ignoring registrator Dogu.
+
 ## [v0.2.0] - 2024-03-20
 ### Added
-- [71] Add optional volume mount for self-signed certificate for the dogu registry.
+- [#71] Add optional volume mount for self-signed certificate for the dogu registry.
 
 ## [v0.1.0] - 2024-03-20
 ### Added